requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
coloredlogs==15.0.1
croniter==5.0.1
cryptography==43.0.3
dataclasses-json==0.6.7
debtcollector==3.0.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dogpile.cache==1.3.3
durationpy==0.9
eventlet==0.37.0
fastapi==0.115.4
fasteners==0.19
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
futurist==3.0.0
google-auth==2.36.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iso8601==2.1.0
Jinja2==3.1.4
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keystoneauth1==5.8.0
keystonemiddleware==10.7.1
kombu==5.4.2
kubernetes==31.0.0
lancedb==0.15.0
langchain==0.3.7
langchain-chroma==0.1.4
langchain-community==0.3.5
langchain-core==0.3.15
langchain-mistralai==0.2.1
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langgraph==0.2.45
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langsmith==0.1.142
legacy-cgi==2.6.1
logutils==0.3.5
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mistral==19.0.0
mistral-lib==3.0.0
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
netaddr==1.3.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.0
openai==1.54.3
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
orjson==3.10.11
os-service-types==1.7.0
oslo.cache==3.8.0
oslo.concurrency==6.1.0
oslo.config==9.6.0
oslo.context==5.6.0
oslo.db==16.0.0
oslo.i18n==6.4.0
oslo.log==6.1.2
oslo.messaging==14.9.1
oslo.metrics==0.9.0
oslo.middleware==6.2.0
oslo.policy==4.5.0
oslo.serialization==5.5.0
oslo.service==3.5.0
oslo.utils==7.4.0
osprofiler==4.2.0
overrides==7.7.0
packaging==24.2
paramiko==3.5.0
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.0
pecan==1.5.1
ply==3.11
posthog==3.7.0
prettytable==3.12.0
prometheus_client==0.21.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycadf==4.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.19.1
PyNaCl==1.5.0
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-keystoneclient==5.5.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
Routes==2.5.1
rpds-py==0.21.0
rsa==4.9
setuptools==75.4.0
shellingham==1.5.4
simplegeneric==0.8.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.2
statsd==4.0.1
stevedore==5.3.0
sympy==1.13.3
tenacity==9.0.0
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
tiktoken==0.8.0
tokenizers==0.20.3
tooz==6.3.0
tqdm==4.67.0
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
voluptuous==0.15.2
watchfiles==0.24.0
wcwidth==0.2.13
WebOb==1.8.9
websocket-client==1.8.0
websockets==14.0
wrapt==1.16.0
WSME==0.12.1
yappi==1.6.9
yaql==3.0.0
yarl==1.17.1
zipp==3.21.0
